requirements.txt
streamlit
opencv-python
numpy
scipy
scikit-image
matplotlib
pandas
Pillow
streamlit-webrtc
